Слит весь код, исправлены ошибки
